polish docker-wipe

+++ docker-wipe.sh
@@ -2,12 +2,17 @@
 #
 # delete all docker container and images
 
-SUDO=sudo
+[ docker version 2> /dev/null ] || SUDO=sudo
 
 #~ FILTER="head -1"
 FILTER="cat"
 
 df -h /
+
+if [ $($SUDO docker images -q -a | wc -l) -eq 0 ] ; then
+	echo "no docker images found. quit." >&2
+	exit 0
+fi
 
 # first kill all which are still running, then delete all containers, then all images
 $SUDO docker ps    --format "$SUDO docker kill '{{.ID}}'" | $FILTER | bash &&
